+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ### Added:
 - Drag click to select items in address list and scratchpad
 
+### Updated:
+- Use simpler and slightly faster `pread()` to read process memory
+
 ### Fixed:
 - Input box for Value +- is usable even when strictly integer types are selected
 - Integer search with Value +- near integer limits works correctly instead of rolling over and giving no results